requirements.txt
nextcord
pymongo
python-dotenv